api/requirements.txt
fastapi
loguru
uvicorn
requests
psycopg2-binary
sqlalchemy